requirements.txt
faiss-cpu
sentence-transformers
fastapi
uvicorn
pydantic>=2
pandas
numpy
nltk
tqdm
PyPDF2
